requirements.txt
torch>=2.0.0
transformers>=4.37.0
datasets>=2.14.0
accelerate>=0.27.0
peft>=0.7.0
trl>=0.7.10
bitsandbytes>=0.41.0
scipy
scikit-learn
sentencepiece
rich>=13.0.0
psutil>=5.9.0
nvidia-ml-py3 